+++ src/dlls/objects/535_VFP_ObjCrea/VFP_ObjCrea.c
@@ -59,7 +59,7 @@ void VFP_ObjCreator_update(GameObject* obj)
     VfpObjCreatorPlacement* placement = (VfpObjCreatorPlacement*)obj->anim.placementData;
     VfpObjCreatorState* state = obj->extra;
 
-    if (Obj_CanSetupObject() == 0)
+    if ((u8)Obj_CanSetupObject() == 0)
     {
         return;
     }
